+++ client-reactive/src/main/java/com/influxdb/client/reactive/QueryReactiveApi.java
@@ -27,12 +27,16 @@ import javax.annotation.concurrent.ThreadSafe;
 
 import com.influxdb.client.InfluxDBClientOptions;
 import com.influxdb.client.domain.Dialect;
+import com.influxdb.client.domain.Query;
 import com.influxdb.query.FluxRecord;
 
 import org.reactivestreams.Publisher;
 
 /**
- * The client that allow perform Flux query against theInfluxDB 2.0by a reactive way.
+ * The client that allow perform Flux query against the InfluxDB 2.0 by a reactive way.
+ *
+ * For parametrized queries use {@link Query} object, see <code>com.influxdb.client.QueryApi</code> in Java module
+ * for more details.
  *
  * @author Jakub Bednar (bednar@github) (21/11/2018 07:19)
  */
@@ -55,12 +59,35 @@ public interface QueryReactiveApi {
      * Returns {@link Publisher} emitting {@link FluxRecord}s which are matched the query.
      * If none found than return empty sequence.
      *
+     * <p>The {@link InfluxDBClientOptions#getOrg()} will be used as source organization.</p>
+     *
+     * @param query the Flux query to execute
+     * @return {@link Publisher} of {@link FluxRecord}s
+     */
+    @Nonnull
+    Publisher<FluxRecord> query(@Nonnull Query query);
+
+    /**
+     * Returns {@link Publisher} emitting {@link FluxRecord}s which are matched the query.
+     * If none found than return empty sequence.
+     *
      * @param query the Flux query to execute
      * @param org   specifies the source organization
      * @return {@link Publisher} of {@link FluxRecord}s
      */
     @Nonnull
     Publisher<FluxRecord> query(@Nonnull final String query, @Nonnull final String org);
+
+    /**
+     * Returns {@link Publisher} emitting {@link FluxRecord}s which are matched the query.
+     * If none found than return empty sequence.
+     *
+     * @param query the Flux query to execute
+     * @param org   specifies the source organization
+     * @return {@link Publisher} of {@link FluxRecord}s
+     */
+    @Nonnull
+    Publisher<FluxRecord> query(@Nonnull final Query query, @Nonnull final String org);
 
     /**
      * Execute a Flux against the Flux service.
@@ -91,6 +118,20 @@ public interface QueryReactiveApi {
                            @Nonnull final Class<M> measurementType);
 
     /**
+     * Execute a Flux against the Flux service.
+     *
+     * @param query           the flux query to execute
+     * @param org             specifies the source organization
+     * @param measurementType the class type used to which will be result mapped
+     * @param <M>             the type of the measurement (POJO)
+     * @return {@link Publisher} emitting a POJO mapped to {@code measurementType} which are matched
+     * the query or empty sequence if none found.
+     */
+    <M> Publisher<M> query(@Nonnull final Query query,
+                           @Nonnull final String org,
+                           @Nonnull final Class<M> measurementType);
+
+    /**
      * Returns {@link Publisher} emitting {@link FluxRecord}s which are matched the query.
      * If none found than return empty sequence.
      *
@@ -112,6 +153,18 @@ public interface QueryReactiveApi {
      */
     @Nonnull
     Publisher<FluxRecord> query(@Nonnull final Publisher<String> queryStream, @Nonnull final String org);
+
+    /**
+     * Returns {@link Publisher} emitting {@link FluxRecord}s which are matched the query.
+     * If none found than return empty sequence.
+     *
+     * @param queryStream the Flux query publisher
+     * @param org         specifies the source organization
+     * @return {@link Publisher} of {@link FluxRecord}s
+     */
+    @Nonnull
+    Publisher<FluxRecord> queryQuery(@Nonnull Publisher<Query> queryStream,
+                                     @Nonnull String org);
 
     /**
      * Returns the {@link Publisher} emitting POJO stream.
@@ -155,6 +208,17 @@ public interface QueryReactiveApi {
      */
     @Nonnull
     Publisher<String> queryRaw(@Nonnull final String query);
+
+    /**
+     * Returns {@link Publisher} emitting raw response fromInfluxDB 2.0server line by line.
+     *
+     * <p>The {@link InfluxDBClientOptions#getOrg()} will be used as source organization.</p>
+     *
+     * @param query the Flux query to execute
+     * @return {@link Publisher} of response lines
+     */
+    @Nonnull
+    Publisher<String> queryRaw(@Nonnull Query query);
 
     /**
      * Returns {@link Publisher} emitting raw response fromInfluxDB 2.0server line by line.
@@ -218,6 +282,20 @@ public interface QueryReactiveApi {
     /**
      * Returns {@link Publisher} emitting queryRaw response fromInfluxDB 2.0server line by line.
      *
+     * @param dialect Dialect is an object defining the options to use when encoding the response.
+     *                <a href="http://bit.ly/flux-dialect">See dialect SPEC.</a>.
+     * @param query   the Flux query to execute
+     * @param org     specifies the source organization
+     * @return {@link Publisher} of response lines
+     */
+    @Nonnull
+    Publisher<String> queryRaw(@Nonnull final Query query,
+                               @Nullable final Dialect dialect,
+                               @Nonnull final String org);
+
+    /**
+     * Returns {@link Publisher} emitting queryRaw response fromInfluxDB 2.0server line by line.
+     *
      * <p>The {@link InfluxDBClientOptions#getOrg()} will be used as source organization.</p>
      *
      * @param dialect     Dialect is an object defining the options to use when encoding the response.
@@ -242,4 +320,18 @@ public interface QueryReactiveApi {
     Publisher<String> queryRaw(@Nonnull final Publisher<String> queryStream,
                                @Nullable final Dialect dialect,
                                @Nonnull final String org);
+
+    /**
+     * Returns {@link Publisher} emitting queryRaw response fromInfluxDB 2.0server line by line.
+     *
+     * @param dialect     Dialect is an object defining the options to use when encoding the response.
+     *                    <a href="http://bit.ly/flux-dialect">See dialect SPEC.</a>.
+     * @param queryStream the Flux query publisher
+     * @param org         specifies the source organization
+     * @return {@link Publisher} of response lines
+     */
+    Publisher<String> queryRawQuery(@Nonnull final Publisher<Query> queryStream,
+                                    @Nullable final Dialect dialect,
+                                    @Nonnull final String org);
+
 }